Fixed deprecated, add class description to package doc
